requirements.txt
requests~=2.32.3
lxml~=5.2.2
python-dotenv~=1.0.1
mysql-connector-python~=9.2.0
pycryptodome~=3.22.0
toml~=0.10.2
python-box~=7.3.2
colorlog~=6.9.0